pyproject.toml
[tool.poetry]
name = "transfinance"
version = "0.1.0"
description = ""
authors = ["erondondron <smirnov.andrey.serg@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
tinvest = "^3.0.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
mypy = "^0.930"
black = "^21.12b0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80
